+++ scripts/uplift_analysis_2025/aero_up.py
