+++ scripts/uplift_analysis_2025/wind_up_config/asset/HOT.yaml
