+++ scripts/uplift_analysis_2025/wind_up_config/asset/turbine_type/SWT_2p3_82.yaml
