+++ scripts/wedowind_challenge_2025/s01_download_reference_data.py
