+++ scripts/wedowind_challenge_2025/s02_generate_datapack.py
