+++ scripts/wedowind_challenge_2025/solution_using_wind_up.py
